initial commit
The parts of the Compose specification completed so far are the
top-level fields `version`, `name`, and `include`, also extensions.
In addition, for `services`, the `attach`, `build`
(short and long form), and `isolation` fields are complete.

Much of the work so far has been developing the supporting
infrastructure, such as types in the `common` module like `ShortOrLong`,
the `serde` module, and the macros in the `compose_spec_macros` crate.
There has also been a lot of parsing/validation work and developing
public helper methods for various field types like `Image`.