fix(service): image registry with port is valid
Image names with a registry that have a port are now valid.

Changed `compose_spec::service::image::Name::new()` to allow for
using `compose_spec::service::Image::set_registry()` with a registry
with a port. The first part of a name is now treated as a registry if it
has a dot (.) regardless of whether the name has multiple parts.

Added `compose_spec::service::image::InvalidNamePart::RegistryPort`
error variant for when a registry's port is not a valid port number.

Refactored image tests to not use `unwrap()`.

Fixes: #22