feat: add variable interpolation in compose files
Introduce the `Options::interpolate_vars()`
function to add name-value mappings that will be
used to replace variable placeholders in YAML
strings when `Compose` is constructed.

The behavior is unchanged if no variables are
given: strings containing placeholders are used
as-is and may lead to errors if they don't pass
validation (which is likely, since they'll contain
constructs like
`${FOO:-${BAR?missing foo & bar}}`
).

If any variables are given to `Options`, the
parser will be invoked and attempt to replace all
of the variables in the YAML source and throw an
error if it is unable to.