volume: check if starting with `/`` instead of using Path::is_absolute
Podman is only running on UNIX-like systems; even on Windows, it is running inside WSLv2.
Thus, its check for absolute path will always only check if said path starts with a `/`.
On non-UNIX platform (e.g. WASM), `Path::is_absolute` will return `false` for paths starting with `/`, which is not what we want.